requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
#
oslo.config>=1.9.3,<1.10.0 # Apache-2.0
oslo.messaging>1.8.0,<1.9.0 # Apache-2.0
oslo.serialization>=1.4.0,<1.5.0 # Apache-2.0
pecan>=1.0.2
python-heatclient>=0.4.0
python-keystoneclient>=1.3.4
python-glanceclient>=0.17.3
werkzeug>=0.11.5
eventlet!=0.17.0,>=0.16.1
oslo.db>=1.7.0,<1.8.0 # Apache-2.0
oslo.log>=1.0.0,<1.1.0 # Apache-2.0
oslo.utils>=1.4.0,<1.5.0 # Apache-2.0
pbr!=0.7,<1.0,>=0.6
SQLAlchemy<=0.9.99,>=0.9.7
sqlalchemy-migrate>=0.9.5
lxml>=2.3
PasteDeploy>=1.5.0
enum34>=0.9.23
